Fix: wrapped useSearchParams in suspense
