+++ src/app/products/page.tsx
@@ -3,6 +3,8 @@ import { getQueryClient } from "@/providers/get-query-client";
 import { getProducts } from "@/utils/api/api";
 
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
+import { Loader } from "lucide-react";
+import { Suspense } from "react";
 
 export default function ProductHome() {
   const queryClient = getQueryClient();
@@ -16,7 +18,9 @@ export default function ProductHome() {
     <section className="flex min-h-[calc(100vh-96px)] w-full mx-auto flex-col mt-24 md:mt-16 ">
       <HydrationBoundary state={dehydrate(queryClient)}>
         {/* <p className="">Page.tsx</p> */}
-        <ProductsPage />
+        <Suspense fallback={<Loader className="w-5 h-5 animate-spin" />}>
+          <ProductsPage />
+        </Suspense>
       </HydrationBoundary>
     </section>
   );
