+++ src/components/navbar/Navbar.tsx
@@ -1,8 +1,8 @@
-import React from "react";
+import React, { Suspense } from "react";
 import { Button } from "../ui/button";
 import CategoriesList from "../catagories-list/CategoriesList";
 import Search from "../search/Search";
-import { Power, ShoppingCart } from "lucide-react";
+import { Loader, Power, ShoppingCart } from "lucide-react";
 import Logo from "./Logo";
 import Link from "next/link";
 import { signIn, signOut } from "../../../auth";
@@ -13,11 +13,15 @@ const Navbar = () => {
     <section className="fixed top-0 flex flex-col md:flex-row gap-1 items-center w-full h-24 md:h-16 bg-primary text-white px-1 md:px-4 z-50">
       <div className="flex items-center justify-between py-2 w-full">
         <div className="flex gap-2">
-          <CategoriesList />
+          <Suspense fallback={<Loader className="w-5 h-5 animate-spin" />}>
+            <CategoriesList />
+          </Suspense>
           <Logo />
         </div>
         <span className="hidden md:block w-full max-w-5xl">
-          <Search placeholder="Search" />
+          <Suspense fallback={<Loader className="w-5 h-5 animate-spin" />}>
+            <Search placeholder="Search" />
+          </Suspense>
         </span>
         <div className="flex items-center gap-3">
           <AuthButtons />
@@ -28,7 +32,9 @@ const Navbar = () => {
         </div>
       </div>
       <div className="w-full md:hidden">
-        <Search placeholder="Search" />
+        <Suspense fallback={<Loader className="w-5 h-5 animate-spin" />}>
+          <Search placeholder="Search" />
+        </Suspense>
       </div>
     </section>
   );
