+++ src/components/productsList/SortOrder.tsx
@@ -9,7 +9,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { urlParam } from "@/hooks/URL_Param";
-import { useRouter, useSearchParams } from "next/navigation";
+import { useRouter } from "next/navigation";
 
 const SortOrder = ({
   selectedCategory,
